third_party/WebKit/LayoutTests/svg/animations/animate-no-matching-end-value-expected.txt

SVG 1.1 dynamic animation tests

Test animate intervals having begin-value without a matching end-value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect1.x.animVal.value is 100
PASS rect1.x.baseVal.value is 0
PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect1.x.animVal.value is 100
PASS rect1.x.baseVal.value is 0
PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
